Made comments.
I've gone through the app casually and made a few notes here and there on
areas of improvement. For the most part, you could do with ramping up
the security of the backend and frontend, and you could lean more into the
react framework to componentize your frontend more as you are repeating
alot of code when you dont need to.